Rewrite function module using std::map
The function module has previously been implemented with std::vector
instead of the more correct std::map which this commit fixes.

m_function's copy constructors has been removed and replaced by move constructors.
The deletion of user functions has been changed to take place in m_function instead of
function_interface. This results in a more clear ownership. As an
result of this the m_function subclass m_function_const was introduced which can be
copied but can not contain user-functions.