Make forward functions create signature strings
The forward functions will after the changes in this commit produce
signature strings. Signature strings are strings which display the
arguments required by a function. For example the signature of sqrt
is "(number)" while the operator '+' has "(number,number)". At the
moment return types are not included in the signature, but this
while be changed in the future. The signature string of a function
call will be included in exception strings produced by the forward
function. To get the signature string from forward functions it is
possible to pass nullptr instead of a valid tree node, which causes
forward to return a char_mat containing the signature.

To get a example of these new signature strings run core_tester and
select menu item #3 to print the function table which will include
the signatures.

This commit also fixes a bug regarding strings. A string like "h i"
would be converted into the string "hi" after evaluation. The white-
space is incorrectly removed by the trim function. The trim function
is now "string aware".