Add optional signature argument names
To extend the features added 4cd17e9 this commit adds the posibility
to define custom argument names to be displayed in the generated
signature string. For exampel "number x" instead of just "number"
The type of the argument will be prepended the name. This features
is utilized for the it function.