Make space characters significant in parsing
Previous version of the parser removed all spaces in the string
before it started parsing. This commit changes this, instead
whitespaces are valid characters that the parser detects. Any
space causes any running parser to stop and then starting a new
one. For example "sqrt x" would normaly be treated as "sqrtx" which
the variable parser would recognize as as a variable named sqrtx, but
with the changes in this commit "sqrt x" would be parsed as two strings
"sqrt" and "x", and "sqrt" would be recognized as a function name. Thus
mathlibra now correctly treats "sqrt x" as the square root of variable
x.

Extra code has also been added to the parser to make parantheses
optional arround function arguments. If left out the parser will
use space characters as delimiters instead. For example
"sqrt x+1 +2" is treated as "sqrt(x+1)+2". Note that the spaces
directly following the function name is ignored.