Make forward function return value templated
It is now possible to specify the return value of the
forward function, thus making it possible to use it directly
on functions returning double, or base_type<T> by value. In this
case forward will will create a copy of it in mathlibras heap. For
plugins this does not mean any extra overhead than today as the returned
type pointer is already being moved to this heap.